Add new navigation app
