+++ wagtailio/core/blocks.py
@@ -654,22 +654,3 @@ class HomePageStoryBlock(blocks.StreamBlock):
 
     class Meta:
         template = "patterns/components/streamfields/home_page_story_block.html"
-
-
-class NavLinkBlock(CTALinkMixin):
-    text = blocks.CharBlock(max_length=100)
-    cta_page = blocks.PageChooserBlock(label="Page", required=False)
-    cta_url = blocks.URLBlock(label="URL", required=False)
-
-    class Meta:
-        icon = "link"
-        label = "Link"
-
-
-class NavSectionBlock(blocks.StructBlock):
-    heading = blocks.CharBlock(max_length=255)
-    links = blocks.ListBlock(NavLinkBlock())
-
-
-class NavStreamField(blocks.StreamBlock):
-    section = NavSectionBlock()
