+++ wagtailio/navigation/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class NavigationConfig(AppConfig):
+    name = "wagtailio.navigation"
+    default_auto_field = "django.db.models.AutoField"
