+++ wagtailio/project_styleguide/templatetags/navigation_tags.py
@@ -1,5 +1,5 @@
 from pattern_library.monkey_utils import override_tag
 
-from wagtailio.core.templatetags.navigation_tags import register
+from wagtailio.navigation.templatetags.navigation_tags import register
 
 override_tag(register, name="footer_navigation")
